+++ src/internal/SincResampler.h
@@ -17,21 +17,21 @@ public:
     // scaleFactor == sourceSampleRate / destinationSampleRate
     // kernelSize can be adjusted for quality (higher is better)
     // numberOfKernelOffsets is used for interpolation and is the number of sub-sample kernel shifts.
-    SincResampler(double scaleFactor, unsigned kernelSize = 32, unsigned numberOfKernelOffsets = 32);
+    SincResampler(double scaleFactor, size_t kernelSize = 32, size_t numberOfKernelOffsets = 32);
     
     // Processes numberOfSourceFrames from source to produce numberOfSourceFrames / scaleFactor frames in destination.
-    void process(const float* source, float* destination, unsigned numberOfSourceFrames);
+    void process(const float* source, float* destination, size_t numberOfSourceFrames);
 
     // Process with input source callback function for streaming applications.
     void process(AudioSourceProvider*, float* destination, size_t framesToProcess);
 
 protected:
     void initializeKernel();
-    void consumeSource(float* buffer, unsigned numberOfSourceFrames);
+    void consumeSource(float* buffer, size_t numberOfSourceFrames);
     
     double m_scaleFactor;
-    unsigned m_kernelSize;
-    unsigned m_numberOfKernelOffsets;
+    size_t m_kernelSize;
+    size_t m_numberOfKernelOffsets;
 
     // m_kernelStorage has m_numberOfKernelOffsets kernels back-to-back, each of size m_kernelSize.
     // The kernel offsets are sub-sample shifts of a windowed sinc() shifted from 0.0 to 1.0 sample.
@@ -42,13 +42,13 @@ protected:
     double m_virtualSourceIndex;
     
     // This is the number of destination frames we generate per processing pass on the buffer.
-    unsigned m_blockSize;
+    size_t m_blockSize;
 
     // Source is copied into this buffer for each processing pass.
     AudioFloatArray m_inputBuffer;
 
     const float* m_source;
-    unsigned m_sourceFramesAvailable;
+    size_t m_sourceFramesAvailable;
     
     // m_sourceProvider is used to provide the audio input stream to the resampler.
     AudioSourceProvider* m_sourceProvider;
